fixed markfile generation bug
